requirements.txt
future
emoji
lxml
beautifulsoup4
requests
sqlalchemy==1.3.20
python-telegram-bot==12.8
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
ffmpeg
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans==3.1.0a0
jikanpy
speedtest-cli
coffeehouse
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
spamwatch
alphabet_detector
pyrate-limiter
cachetools
telethon==1.16.4
youtube_dl
hachoir
asyncio
gtts
bs4
tswift
pyrogram
tgcrypto
